clean role
