+++ playground/app/features/roles/components/__init__.py
@@ -2,7 +2,6 @@ from app.features.roles.components.commons import permission_checkbox_item
 from app.features.roles.components.role_create_form import role_create_form
 from app.features.roles.components.role_update_form import role_update_form
 from app.features.roles.components.roles_header import roles_header
-from app.features.roles.components.roles_limits import roles_limits
 from app.features.roles.components.roles_list import roles_list
 from app.features.roles.components.roles_pagination import roles_pagination
 
@@ -11,7 +10,6 @@ __all__ = [
     "role_create_form",
     "roles_header",
     "roles_list",
-    "roles_limits",
     "role_update_form",
     "roles_pagination",
 ]
